requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.33.0
colorama==0.4.6
dj-database-url==2.0.0
Django
django-cloudinary-storage==0.3.0
django-cors-headers==4.1.0
django-heroku==0.3.1
djangorestframework==3.14.0
ffmpeg-python==0.2.0
filelock==3.12.2
frozenlist==1.3.3
future==0.18.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.0
llvmlite
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx
nltk==3.8.1
numba
numpy
openai==0.27.8
openai-whisper==20230124
psycopg2==2.9.6
python-dotenv
pytz==2023.3
regex==2023.6.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy
torch
tqdm
typing_extensions
tzdata
urllib3==1.26.16
whitenoise
yarl
gunicorn
httpx
pydantic
websockets